repodata/backend/requirements.txt
fastapi
uvicorn
requests
pydantic
python-multipart
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
